+++ madblog/app.py
@@ -15,6 +15,7 @@ from .config import config
 from .feeds import FeedAuthor, FeedParser
 from .latex import MarkdownLatex
 from .mermaid import MarkdownMermaid
+from .toc import MarkdownTocMarkers
 from .notifications import SmtpConfig, build_webmention_email_notifier
 from .storage.mentions import FileWebmentionsStorage
 from ._sorters import PagesSorter, PagesSortByTime
@@ -270,7 +271,17 @@ class BlogApp(Flask):
             published=metadata["published"].strftime("%b %d, %Y"),
             content=markdown(
                 content,
-                extensions=["fenced_code", "codehilite", "tables", MarkdownLatex(), MarkdownMermaid()],
+                extensions=[
+                    "fenced_code",
+                    "codehilite",
+                    "tables",
+                    "toc",
+                    "attr_list",
+                    "sane_lists",
+                    MarkdownTocMarkers(),
+                    MarkdownLatex(),
+                    MarkdownMermaid(),
+                ],
             ),
             skip_header=skip_header,
             skip_html_head=skip_html_head,